moving compute stats to shared utils; adjusting token budget target
