+++ datasets_preparation/data_preparation_utils.py
@@ -588,3 +588,38 @@ def token_budget_dataset_mix(*, datasets, weights, target_tokens, seed):
     prepared_dataset = concatenate_datasets(prepared_datasets)
 
     return prepared_dataset
+
+def compute_stats(prepared_dataset):
+    stats_per_dataset = {}
+
+    for source, total_tokens, supervised_tokens in zip(
+        prepared_dataset['source'],
+        prepared_dataset['total_tokens'],
+        prepared_dataset['supervised_tokens']
+    ):
+        if source not in stats_per_dataset:
+            stats_per_dataset[source] = {
+                'examples': 0,
+                'total_tokens': 0,
+                'supervised_tokens': 0,
+            }
+
+        stats_per_dataset[source]['examples'] += 1
+        stats_per_dataset[source]['total_tokens'] += total_tokens
+        stats_per_dataset[source]['supervised_tokens'] += supervised_tokens
+
+    total_examples = sum(s['examples'] for s in stats_per_dataset.values())
+    total_tokens = sum(s['total_tokens'] for s in stats_per_dataset.values())
+    total_supervised_tokens = sum(s['supervised_tokens'] for s in stats_per_dataset.values())
+
+    for source, source_stats in stats_per_dataset.items():
+        stats_per_dataset[source]['examples %'] = f'{source_stats["examples"] / total_examples:.2%}'
+        stats_per_dataset[source]['tokens %'] = f'{source_stats["total_tokens"] / total_tokens:.2%}'
+        stats_per_dataset[source]['supervised_tokens %'] = f'{source_stats["supervised_tokens"] / total_supervised_tokens:.2%}'
+
+    return {
+        'total_examples': total_examples,
+        'total_tokens': total_tokens,
+        'total_supervised_tokens': total_supervised_tokens,
+        'sources': stats_per_dataset
+    }
